requirements.txt
# Merged requirements for all Sofia services

# Agent Service Dependencies
langchain>=0.1.0
langchain-core>=0.3.51,<0.4.0
langchain-openai>=0.0.5
langgraph>=0.0.25
httpx>=0.27.0
httpx_sse>=0.3.1
pydantic>=2.7.4,<3.0.0
python-dotenv==1.0.0
asyncio==3.4.3
fastapi>=0.95.1
uvicorn>=0.22.0
openai>=1.0.0
starlette>=0.27.0
sse_starlette>=2.2.0 